+++ lib/plausible/google/ga4/report_request.ex
@@ -3,6 +3,13 @@ defmodule Plausible.Google.GA4.ReportRequest do
   Report request struct for Google Analytics 4 API
   """
 
+  @excluded_event_names [
+    "page_view",
+    "session_start",
+    "first_visit",
+    "user_engagement"
+  ]
+
   defstruct [
     :dataset,
     :dimensions,
@@ -11,6 +18,7 @@ defmodule Plausible.Google.GA4.ReportRequest do
     :property,
     :access_token,
     :offset,
+    :dimension_filter,
     :limit
   ]
 
@@ -91,6 +99,25 @@ defmodule Plausible.Google.GA4.ReportRequest do
       #     "bounces = sessions - engagedSessions"
       #   ]
       # },
+      %__MODULE__{
+        dataset: "imported_custom_events",
+        dimensions: ["date", "eventName", "linkUrl"],
+        metrics: [
+          "totalUsers",
+          "eventCount"
+        ],
+        dimension_filter: %{
+          "notExpression" => %{
+            "filter" => %{
+              "fieldName" => "eventName",
+              "inListFilter" => %{
+                "values" => @excluded_event_names,
+                "caseSensitive" => true
+              }
+            }
+          }
+        }
+      },
       %__MODULE__{
         dataset: "imported_locations",
         dimensions: ["date", "countryId", "region", "city"],